+++ src/main/java/com/complete/todayspace/domain/user/repository/UserRefreshTokenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RefreshTokenRepository extends CrudRepository<UserRefreshToken, Long> {
+public interface UserRefreshTokenRepository extends CrudRepository<UserRefreshToken, Long> {
 }
